Nieuwe testen voor rules
