+++ JavaServer/calzone/src/main/java/com/vub/scheduler/SchedulerScoreRules.drl
@@ -196,12 +196,12 @@ end
 */
 rule studentAgendaDurationViolated
 	when
-		$traject : Traject()
-		and
-		Date($day := day, $month := month, $year := year)
-		and
-		accumulate(Entry($day := startingDate.day, $month := startingDate.month, $year := startingDate.year, 
-						$duration : courseComponent.duration, courseComponent.course memberOf $traject.courses);
+		//$traject : Traject()
+		//and
+		//Date($day := getDay(), $month := getMonth(), $year := getYear())
+		//and
+		accumulate(Entry($day := startingDate.getDay(), $month := startingDate.getMonth(), $year := startingDate.getYear(), 
+						$duration : courseComponent.getDuration()); //, courseComponent.course memberOf $traject.courses);
 			       $totalDuration : sum($duration);
 			       $totalDuration > 9)
 	then
@@ -223,10 +223,10 @@ end
 rule spareHoursViolated
 	when
 		$traject : Traject()
-		Date($day : day, $month : month, $year : year) 
+		Date($day := getDay(), $month := getMonth(), $year := getYear()) 
 		$orderedLecturesOnSameDay: ArrayList() from collect 
-			(Entry($day := startingDate.day, $month := startingDate.month,
-				$year := startingDate.year, courseComponent.course memberOf $traject.courses))
+			(Entry($day := startingDate.getDay(), $month := startingDate.getMonth(),
+				$year := startingDate.getYear(), courseComponent.course memberOf $traject.courses))
 	then
 		sort($orderedLecturesOnSameDay);
 		scoreHolder.addSoftConstraintMatch(kcontext, checkSpareHoursAndNoonBreak($orderedLecturesOnSameDay));
