--- JavaServer/calzone/src/main/java/com/vub/scheduler/TrackingAgendaEventListener.java
@@ -1,69 +1,0 @@
-package com.vub.scheduler;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
-
-import org.kie.api.definition.rule.Rule;
-import org.kie.api.event.rule.AfterMatchFiredEvent;
-import org.kie.api.event.rule.DefaultAgendaEventListener;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public class TrackingAgendaEventListener extends DefaultAgendaEventListener {
-    final Logger logger = LoggerFactory.getLogger(getClass());
-
-
-    private List<Activation> activationList = new ArrayList<Activation>();
-
-    @Override
-    public void afterMatchFired(AfterMatchFiredEvent event) {
-        Rule rule = event.getMatch().getRule();
-
-        String ruleName = rule.getName();
-        Map<String, Object> ruleMetaDataMap = rule.getMetaData();
-
-        activationList.add(new Activation(ruleName));
-        StringBuilder sb = new StringBuilder("Rule fired: " + ruleName);
-
-        if (ruleMetaDataMap.size() > 0) {
-            sb.append("\n  With [" + ruleMetaDataMap.size() + "] meta-data:");
-            for (String key : ruleMetaDataMap.keySet()) {
-                sb.append("\n    key=" + key + ", value="
-                        + ruleMetaDataMap.get(key));
-            }
-        }
-
-        logger.debug(sb.toString());
-    }
-
-    public boolean isRuleFired(String ruleName) {
-        for (Activation a : activationList) {
-            if (a.getRuleName().equals(ruleName)) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    public void reset() {
-        activationList.clear();
-    }
-
-    public final List<Activation> getActivationList() {
-        return activationList;
-    }
-
-    public String activationsToString() {
-        if (activationList.size() == 0) {
-            return "No activations occurred.";
-        } else {
-            StringBuilder sb = new StringBuilder("Activations: ");
-            for (Activation activation : activationList) {
-                sb.append("\n  rule: ").append(activation.getRuleName());
-            }
-            return sb.toString();
-        }
-    }
-
-}
