+++ JavaServer/calzone/src/main/test/com/vub/scheduler/AjacentCourseComponentCVTest.java
@@ -1,0 +1,12 @@
+/**
+ * 
+ */
+package com.vub.scheduler;
+
+/**
+ * @author pieter
+ *
+ */
+public class AjacentCourseComponentCVTest {
+
+}
