+++ JavaServer/calzone/src/main/test/com/vub/scheduler/ConstraintViolationTest.java
@@ -1,0 +1,13 @@
+/**
+ * 
+ */
+package com.vub.scheduler;
+
+
+/**
+ * @author pieter
+ *
+ */
+public class ConstraintViolationTest extends SchedulerTest {
+
+}
